@@ -9,23 +9,23 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@rainbow-me/rainbowkit": "^0.4.6",
+    "@rainbow-me/rainbowkit": "0.7.1",
     "@stitches/react": "1.2.8",
     "next": "12.3.1",
-    "next-axiom": "^0.14.0",
-    "next-themes": "^0.2.0",
+    "next-axiom": "0.14.0",
+    "next-themes": "0.2.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "swr": "^1.3.0",
-    "wagmi": "^0.6.2"
+    "swr": "1.3.0",
+    "wagmi": "0.6.8"
   },
   "devDependencies": {
-    "@types/node": "18.6.2",
-    "@types/react": "18.0.15",
-    "eslint": "8.20.0",
-    "eslint-config-next": "12.2.5",
+    "@types/node": "18.8.3",
+    "@types/react": "18.0.21",
+    "eslint": "8.21.0",
+    "eslint-config-next": "12.3.1",
     "next-compose-plugins": "2.2.1",
     "next-transpile-modules": "9.0.0",
-    "typescript": "4.7.4"
+    "typescript": "4.8.4"
   }
 }
