@@ -9,23 +9,23 @@
     "@radix-ui/react-dropdown-menu": "1.0.0",
     "@radix-ui/react-icons": "1.1.1",
     "@radix-ui/react-switch": "1.0.0",
-    "@rainbow-me/rainbowkit": "^0.5.0",
-    "@sanity/client": "3.3.3",
-    "@sanity/image-url": "^1.0.1",
+    "@rainbow-me/rainbowkit": "0.7.1",
+    "@sanity/client": "3.4.1",
+    "@sanity/image-url": "1.0.1",
     "@stitches/react": "1.2.8",
-    "@zeit/swr": "0.1.2",
-    "ethers": "^5.7.0",
+    "swr": "0.1.2",
+    "ethers": "5.7.1",
     "groq": "2.29.3",
     "modern-normalize": "1.1.0",
     "next": "12.3.1",
-    "next-axiom": "^0.14.0",
+    "next-axiom": "0.14.0",
     "next-themes": "0.2.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-is": "18.2.0",
-    "sass": "1.54.4",
-    "sharp": "0.30.6",
-    "wagmi": "0.6.4"
+    "sass": "1.55.0",
+    "sharp": "0.31.1",
+    "wagmi": "0.6.8"
   },
   "scripts": {
     "dev": "next dev",
@@ -34,12 +34,12 @@
     "lint": "next lint"
   },
   "devDependencies": {
-    "@types/node": "18.7.2",
-    "@types/react": "18.0.17",
-    "eslint": "^8.0.0",
-    "eslint-config-next": "12.2.5",
+    "@types/node": "18.8.3",
+    "@types/react": "18.0.21",
+    "eslint": "8.21.0",
+    "eslint-config-next": "12.3.1",
     "next-compose-plugins": "2.2.1",
     "next-transpile-modules": "9.0.0",
-    "typescript": "4.7.4"
+    "typescript": "4.8.4"
   }
 }
