@@ -32,6 +32,6 @@
     "eslint-config-next": "12.2.5",
     "next-compose-plugins": "^2.2.1",
     "next-transpile-modules": "^9.0.0",
-    "typescript": "4.7.4"
+    "typescript": "4.8.4"
   }
 }
