@@ -1,43 +1,43 @@
 import '../styles/global-styles.scss'
 import '@rainbow-me/rainbowkit/styles.css'
-import {getDefaultWallets, RainbowKitProvider} from '@rainbow-me/rainbowkit'
-import {chain, configureChains, createClient, WagmiConfig} from 'wagmi'
-import {alchemyProvider} from 'wagmi/providers/alchemy'
-import {publicProvider} from 'wagmi/providers/public'
-import {ThemeProvider} from 'next-themes'
-import type {AppProps} from 'next/app'
+import { getDefaultWallets, RainbowKitProvider } from '@rainbow-me/rainbowkit'
+import { chain, configureChains, createClient, WagmiConfig } from 'wagmi'
+import { alchemyProvider } from 'wagmi/providers/alchemy'
+import { publicProvider } from 'wagmi/providers/public'
+import { ThemeProvider } from 'next-themes'
+import type { AppProps } from 'next/app'
 import Frame from '@/components/Frame'
 
-export {reportWebVitals} from 'next-axiom';
+export { reportWebVitals } from 'next-axiom'
 
-const {chains, provider} = configureChains(
-    [chain.mainnet, chain.polygon, chain.optimism, chain.arbitrum],
-    [alchemyProvider({apiKey: process.env.ALCHEMY_ID}), publicProvider()],
+const { chains, provider } = configureChains(
+  [chain.mainnet, chain.polygon, chain.optimism, chain.arbitrum],
+  [alchemyProvider({ apiKey: process.env.ALCHEMY_ID }), publicProvider()],
 )
 
-const {connectors} = getDefaultWallets({
-    appName: '0xbp.io Wallet Visualizer',
-    chains,
+const { connectors } = getDefaultWallets({
+  appName: '0xbp.io Wallet Visualizer',
+  chains,
 })
 
 const wagmiClient = createClient({
-    autoConnect: true,
-    connectors,
-    provider,
+  autoConnect: true,
+  connectors,
+  provider,
 })
 
-function MyApp({Component, pageProps}: AppProps) {
-    return (
-        <WagmiConfig client={wagmiClient}>
-            <RainbowKitProvider chains={chains}>
-                <ThemeProvider>
-                    <Frame>
-                        <Component {...pageProps} />
-                    </Frame>
-                </ThemeProvider>
-            </RainbowKitProvider>
-        </WagmiConfig>
-    )
+const MyApp = ({ Component, pageProps }: AppProps) => {
+  return (
+    <WagmiConfig client={wagmiClient}>
+      <RainbowKitProvider chains={chains}>
+        <ThemeProvider>
+          <Frame>
+            <Component {...pageProps} />
+          </Frame>
+        </ThemeProvider>
+      </RainbowKitProvider>
+    </WagmiConfig>
+  )
 }
 
 export default MyApp
