@@ -1,24 +1,23 @@
-import type {AppProps} from 'next/app'
-import {AnimatePresence} from 'framer-motion'
-
-export {reportWebVitals} from 'next-axiom';
+import '@/styles/_global-styles.scss'
+import type { AppProps } from 'next/app'
+import { AnimatePresence } from 'framer-motion'
 
 import Layout from '@/components/layout'
 
-import '@/styles/_global-styles.scss'
+export { reportWebVitals } from 'next-axiom'
 
-const MyApp = ({Component, pageProps, router}: AppProps) => {
-    return (
-        <Layout>
-            <AnimatePresence
-                exitBeforeEnter
-                initial={false}
-                onExitComplete={() => window.scrollTo(0, 0)}
-            >
-                <Component {...pageProps} key={router.route}/>
-            </AnimatePresence>
-        </Layout>
-    )
+const MyApp = ({ Component, pageProps, router }: AppProps) => {
+  return (
+    <Layout>
+      <AnimatePresence
+        exitBeforeEnter
+        initial={false}
+        onExitComplete={() => window.scrollTo(0, 0)}
+      >
+        <Component {...pageProps} key={router.route} />
+      </AnimatePresence>
+    </Layout>
+  )
 }
 
 export default MyApp
