@@ -1,3 +1,3 @@
 import { FC } from 'react'
 
-export type HeaderTypes = FC<{}>
+export type HeaderTypes = FC
