@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useMemo } from 'react'
 import { reducer, initialState } from './reducer'
 import { UserProviderTypes } from './user-context.types'
 
@@ -9,9 +9,13 @@ export const UserContext = React.createContext<any>({
 
 export const UserProvider: UserProviderTypes = ({ children }) => {
   const [state, dispatch] = React.useReducer(reducer, initialState)
+  const userProviderValue = useMemo(
+    () => ({ state, dispatch }),
+    [state, dispatch],
+  )
 
   return (
-    <UserContext.Provider value={[state, dispatch]}>
+    <UserContext.Provider value={userProviderValue}>
       {children}
     </UserContext.Provider>
   )
