@@ -1,6 +1,6 @@
 import '../styles/global-styles.scss'
-import {ThemeProvider} from 'next-themes'
-import type {AppProps} from 'next/app'
+import { ThemeProvider } from 'next-themes'
+import type { AppProps } from 'next/app'
 import React from 'react'
 import Head from 'next/head'
 
@@ -10,76 +10,75 @@ import Footer from '@/components/Footer'
 
 import Constants from '@/config/constants'
 
-import {UserProvider} from '@/contexts/User'
+import { UserProvider } from '@/contexts/User'
 // import { darkTheme } from 'stitches.config'
 
 // Wallet SSO Dependencies
 import '@rainbow-me/rainbowkit/styles.css'
 import {
-    getDefaultWallets,
-    RainbowKitProvider,
-    darkTheme,
-    lightTheme,
+  getDefaultWallets,
+  RainbowKitProvider,
+  darkTheme,
+  lightTheme,
 } from '@rainbow-me/rainbowkit'
-import {chain, configureChains, createClient, WagmiConfig} from 'wagmi'
+import { chain, configureChains, createClient, WagmiConfig } from 'wagmi'
 
-import {alchemyProvider} from 'wagmi/providers/alchemy'
-import {infuraProvider} from 'wagmi/providers/infura'
-import {publicProvider} from 'wagmi/providers/public'
+import { alchemyProvider } from 'wagmi/providers/alchemy'
+import { infuraProvider } from 'wagmi/providers/infura'
+import { publicProvider } from 'wagmi/providers/public'
 
-const {withAxiom} = require('next-axiom');
+export { reportWebVitals } from 'next-axiom'
 
-
-const {chains, provider} = configureChains(
-    [chain.mainnet, chain.polygon, chain.optimism, chain.arbitrum],
-    [
-        alchemyProvider({apiKey: process.env.NEXT_PUBLIC_ALCHEMY_ID}),
-        infuraProvider({apiKey: process.env.NEXT_PUBLIC_INFURA_ID}),
-        publicProvider(),
-    ],
+const { chains, provider } = configureChains(
+  [chain.mainnet, chain.polygon, chain.optimism, chain.arbitrum],
+  [
+    alchemyProvider({ apiKey: process.env.NEXT_PUBLIC_ALCHEMY_ID }),
+    infuraProvider({ apiKey: process.env.NEXT_PUBLIC_INFURA_ID }),
+    publicProvider(),
+  ],
 )
-const {connectors} = getDefaultWallets({
-    appName: Constants.APP_TITLE,
-    chains,
+const { connectors } = getDefaultWallets({
+  appName: Constants.APP_TITLE,
+  chains,
 })
 
 const wagmiClient = createClient({
-    autoConnect: true,
-    connectors,
-    provider,
+  autoConnect: true,
+  connectors,
+  provider,
 })
 
-function MyApp({Component, pageProps}: AppProps) {
-    return (
-        <WagmiConfig client={wagmiClient}>
-            <RainbowKitProvider
-                chains={chains}
-                appInfo={{
-                    appName: Constants.APP_TITLE,
-                }}
-            >
-                <ThemeProvider>
-                    <UserProvider>
-                        <Frame>
-                            <Head>
-                                <meta
-                                    name="viewport"
-                                    content="initial-scale=1.0, maximum-scale=1.0, width=device-width"
-                                />
-                                <title>{Constants.APP_TITLE}</title>
-                            </Head>
+const MyApp = ({ Component, pageProps }: AppProps) => {
+  return (
+    <WagmiConfig client={wagmiClient}>
+      <RainbowKitProvider
+        chains={chains}
+        appInfo={{
+          appName: Constants.APP_TITLE,
+        }}
+      >
+        <ThemeProvider>
+          <UserProvider>
+            <Frame>
+              <Head>
+                <meta
+                  name="viewport"
+                  content="initial-scale=1.0, maximum-scale=1.0, width=device-width"
+                />
+                <title>{Constants.APP_TITLE}</title>
+              </Head>
 
-                            <Header/>
+              <Header />
 
-                            <Component {...pageProps} />
+              <Component {...pageProps} />
 
-                            <Footer/>
-                        </Frame>
-                    </UserProvider>
-                </ThemeProvider>
-            </RainbowKitProvider>
-        </WagmiConfig>
-    )
+              <Footer />
+            </Frame>
+          </UserProvider>
+        </ThemeProvider>
+      </RainbowKitProvider>
+    </WagmiConfig>
+  )
 }
 
 export default MyApp
